+++ src/adapter/admin/index.mjs
@@ -8,8 +8,6 @@ export { transaction } from './transaction.mjs'
 
 export { groups } from './groups.mjs'
 
-export { retry } from './../../retry/index.mjs'
-
 export * from './../../helpers/index.mjs'
 
 export function schema(getSchema) {
