+++ src/adapter/web/index.mjs
@@ -34,8 +34,6 @@ export { groups } from './groups.mjs'
 
 export { transaction } from './transaction.mjs'
 
-export { retry } from './../../retry/index.mjs'
-
 export * from './../../helpers/index.mjs'
 
 export function schema(getSchema) {
