+++ src/helpers/index.d.ts
@@ -1,3 +1,5 @@
 import { TypesaurusHelpers as Helpers } from '../types/helpers'
 
+export const retry: Helpers.RetryFunction
+
 export const silence: Helpers.SilenceFunction
