+++ src/index.d.ts
@@ -104,6 +104,4 @@ export { groups } from './groups'
 
 export { batch } from './batch'
 
-export { retry } from './retry'
-
 export * from './helpers'
