--- src/retry/index.d.ts
@@ -1,3 +1,0 @@
-import { TypesaurusRetry as Retry } from '../types/retry'
-
-export const retry: Retry.Function
